fix: consult @parent_model for association resource name
I have a series of resources whose names don't match their activerecord
model (model_class).  Where normally you'd expect "OrderResource" to map
to the "Order" activerecord class.  I have that case, but I also have
resources like "OrderAwaitingPaymentResource", and
"OrderAwaitingFulfillmentResource".  Each of these Resource classes has
a model_class set to a ruby class that inherits from the "Order" model
class.  E.g.:

    # All normal here:

    class Order < ApplicationRecord
      has_many :line_items
    end

    class OrderResource < Avo::BaseResource
      field :line_items, as: :has_many
    end

    # Note the order policy dissallows attaching line items
    class OrderPolicy
      def attach_line_items?
        false
      end
    end

    # Here's a new model class that inherits from the "real" model class
    # Importantly, it defers to the OrderPolicy

    class Order::AwaitingPayment < Order
      def self.policy_class = OrderPolicy
    end

    # And here's a corresponding resource class that points to the model
    # it should use
    class OrderAwaitingPaymentResource < Avo::BaseResource
      self.model_class = Order::AwaitingPayment
    end

Before this change, if you visited the Orders resource, you'd see line
items, but not be permitted to attach them - as expected according to
the policy.  HOWEVER - you would be able to attach them when viewing the
OrderAwaitingPayment resource.

The problem was that the authorization was happening with the wrong
(unhhydrated) resource.  In the problem scenario,

@reflection.active_record.name == "Order"

but we needed "Order::AwaitingPayment" - which just so happens to be the
parent model in this scenario.

After this change, the correct resource is consulted for authorization.